backend/requirements.txt
flask
flask-cors
python-dotenv
openai
mcp
httpx
